Rename package to prevent conflicts
