+++ android/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.oblongmana.webviewfileuploadandroid">
+    package="com.appfolio.react.webview">
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
 
     <application>
