+++ example/android/app/src/main/java/com/advancedandroidwebviewexample/MainApplication.java
@@ -7,7 +7,7 @@ import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.shell.MainReactPackage;
 import com.facebook.soloader.SoLoader;
-import com.oblongmana.webviewfileuploadandroid.AndroidWebViewPackage;
+import com.appfolio.react.webview.AndroidWebViewPackage;
 
 import java.util.Arrays;
 import java.util.List;
